Renamed functions to api and seperated matches Renamed the functions folder to api for better clarity since this project will likely grow in size to include much more stuff. Also moved match_game and created a match_timeline file to keep match related stuff together and keep the timeline seperate from the main match endpoint.
